+++ lib/gdal/raster_band.rb
@@ -221,7 +221,7 @@ module GDAL
 
     # @return [Boolean]
     def create_mask_band(*flags)
-      flag_value = flags.each_with_object(0) do |flag, result|
+      flag_value = flags.each_with_object(0) do |flag, _result|
         case flag
         when :GMF_ALL_VALID then 0x01
         when :GMF_PER_DATASET then 0x02