+++ lib/ogr/geometries/line_string.rb
@@ -5,12 +5,8 @@ module OGR
     include OGR::Geometry
     include GeometryTypes::Curve
 
-    def self.approximate_arc_angles(center_x, center_y,
-                                    z,
-                                    primary_radius, secondary_radius,
-                                    rotation,
-                                    start_angle, end_angle,
-                                    max_angle_step_size_degrees = 0)
+    def self.approximate_arc_angles(center_x, center_y, z, primary_radius, secondary_radius,
+      rotation, start_angle, end_angle, max_angle_step_size_degrees = 0)
       geometry_ptr = FFI::GDAL::GDAL.OGR_G_ApproximateArcAngles(
         center_x,
         center_y,