+++ History.md
@@ -60,6 +60,8 @@ Format for this file derived from [http://keepachangelog.com](http://keepachange
 * Allow `GDAL::RasterBand#create_mask_band` to take a single flag or many.
 * Allow `GDAL::Dataset`s to be open in shared mode vs non-shared mode. All
   Datasets now default to use shared mode.
+* Allow `GDAL::Driver#create_dataset` to take a block, yielding the dataset
+  then closing it afterwards.
 
 #### OGR
 
@@ -96,6 +98,8 @@ Format for this file derived from [http://keepachangelog.com](http://keepachange
   and `point_geometries`. Since `#point` implies it returns a Point object, but
   the OGR API's related method returns point values (x, y, z), it seemed like
   it would be useful to have a method that returned a geometry.
+* `OGR::DataSource.open` can now take a block, yielding the data source, then
+  closing it afterwards.
 
 ### Bug Fixes